feat: 즐겨찾기 삭제 기능

+++ src/main/java/nextstep/subway/favorite/application/FavoriteService.java
@@ -1,6 +1,7 @@
 package nextstep.subway.favorite.application;
 
 import nextstep.subway.auth.domain.LoginMember;
+import nextstep.subway.exception.FavoriteNotFoundException;
 import nextstep.subway.favorite.domain.Favorite;
 import nextstep.subway.favorite.domain.FavoriteRepository;
 import nextstep.subway.favorite.dto.FavoriteRequest;
@@ -42,7 +43,10 @@ public class FavoriteService {
                 .collect(Collectors.toList());
     }
 
-    public void deleteFavoriteById(Long id) {
-        favoriteRepository.deleteById(id);
+    public void deleteFavorite(LoginMember loginMember, Long id) {
+        Favorite favorite = favoriteRepository.findById(id).orElseThrow(FavoriteNotFoundException::new);
+        if (favorite.isRegisteredBy(loginMember)) {
+            favoriteRepository.deleteById(id);
+        }
     }
 }
